steps: Factor out the shared HTTP/FTP server code
The shared code was copy pasted, factor it out into a ServerContext
class the two servers can inherit from.
